Skip existing

+++ src/main.rs
@@ -56,6 +56,7 @@ async fn main() {
 	// (i.e., 10 requests per second on average)."
 	for (name, set, out_path) in &cards {
 		println!("Downloading: |{}|", name);
+		if out_path.exists() { continue; }
 		let img = get_card_image(&name, set).await.unwrap();
 		std::fs::write(&out_path, img).unwrap();
 		sleep(std::time::Duration::from_millis(150)).await;
